requirements.txt
agno>=1.0.0
groq>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
pymupdf>=1.23.0
python-docx>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
